% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{agscore3}
\alias{agscore3}
\title{agscore3}
\usage{
agscore3(tstart, tstop, event, covar, strata, score, weights, sort1, method2)
}
\arguments{
\item{tstart}{start time}

\item{tstop}{stop time}

\item{event}{event}

\item{strata}{unique non-negative for each stratum}

\item{score}{the vector of subject scores, i.e., exp(beta*z)}

\item{weights}{case weights}

\item{sort1}{sort order of y\link{,1}, within strata}

\item{covar2}{the matrix of covariates}

\item{method}{==1 for efron approx}
}
\description{
Calculate score residuals for a mixed effect cox model. Only calculates for fixed effects.
}
