% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_derivatives.R
\name{Bb}
\alias{Bb}
\title{Wang 2019 likelihood differentiated twice with respect to the fixed effects, \eqn{\beta}.}
\usage{
Bb(parms, X, t, cluster, dij, data)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{data}{tibble with all these columns}

\item{Z}{character vector containing names of Z columns in data.}

\item{wi}{column in data containing cluster-level weights.}

\item{wji}{column in data containing within-cluster weights.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}
}
\description{
Wang 2019 likelihood differentiated twice with respect to the fixed effects, \eqn{\beta}.
}
