% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_derivatives.R
\name{bb2}
\alias{bb2}
\title{This is the K prime prime wrt b from appendix 1 of Ripatti Palmgren. It's a bit different to the b*b block from the hessian.}
\usage{
bb2(parms, X, t, cluster, dij, data, theta, return_matrix = FALSE, ...)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{data}{tibble with all these columns}

\item{Z}{character vector containing names of Z columns in data.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}
}
\description{
This is the K prime prime wrt b from appendix 1 of Ripatti Palmgren. It's a bit different to the b*b block from the hessian.
}
