% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_calc_S0_S1X}
\alias{C_calc_S0_S1X}
\title{C_calc_S0_S1X}
\usage{
C_calc_S0_S1X(time_start, time_stop, exp_risk_score, X)
}
\arguments{
\item{time_start}{start time for an observation}

\item{time_stop}{stop time for an observation}

\item{X}{matrix of covariate data}
}
\description{
Calculate S0 and S1X for a mixed effect cox model. Only calculates for fixed effects.
}
