% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_calc_ui}
\alias{C_calc_ui}
\title{C_calc_ui}
\usage{
C_calc_ui(
  time_start,
  time_stop,
  stat,
  weights,
  exp_risk_score,
  S0,
  S1_X,
  X,
  weighted
)
}
\arguments{
\item{time_start}{start time for an observation}

\item{time_stop}{stop time for an observation}

\item{stat}{indicator, 1 for event, 0 for censoring}

\item{weights}{observation-level sampling weights}

\item{exp_risk_score}{exp(betaX + bZ)}

\item{S0}{sum of exp_risk_score for observations at risk at time_stop}

\item{S1_X}{sum of X * exp_risk_score for  observations at risk at time_stop}

\item{X}{matrix of covariate data}

\item{weighted}{logical scalar indicating if the residuals should be weighted}
}
\description{
Calculate score residuals for a mixed effect cox model. Only calculates for fixed effects.
}
