% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_draw_event_times}
\alias{C_draw_event_times}
\title{C_draw_event_times}
\usage{
C_draw_event_times(
  id,
  start_time,
  end_time,
  status,
  X,
  risk_score,
  baseline_hazard,
  baseline_hazard_start,
  baseline_hazard_end,
  end_of_follow_up,
  origin,
  single
)
}
\arguments{
\item{subject_id}{unique identifier for subject}
}
\description{
Draw event times times using possibly time-varying covariate data
}
