% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_rpexp}
\alias{C_rpexp}
\title{C_rpexp}
\usage{
C_rpexp(n, rate, t, start)
}
\arguments{
\item{rate}{vector of event rates}

\item{t}{vector same length as rate, giving the times the rate changes. The values of t should be ascending order.}

\item{start}{numeric scalar; delayed entry time. The random deviates will be left truncated.}
}
\description{
Draw a single random number from a piecewise exponential distribution.
See msm::rpexp()
}
