% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{calcCrossProducts}
\alias{calcCrossProducts}
\title{Get the cross products}
\usage{
calcCrossProducts(data, f1, f2, n1, n2)
}
\arguments{
\item{data}{a tibble containing the numeric variables to cross multiply}

\item{f1}{first set of variables.}

\item{f2}{second set of variables (can be the same variables as in f1)}

\item{n1}{name for the first set of variables}

\item{n2}{name for the second set of variables.}
}
\description{
Calculate cross products for a set of numeric variables.
}
