% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{calcLinearPredictor}
\alias{calcLinearPredictor}
\title{calculate (XB + Zb) and exp(XB + Zb)}
\usage{
calcLinearPredictor(data, X, Z, parms)
}
\arguments{
\item{data}{a tibble containing X, Z. It can have other columns too.}

\item{X}{character vector of X columns, i.e. fixed-effect covariates}

\item{Z}{character vector of Z columns, i.e. random-effect covariates. Currently this is a set of indicator variables describing cluster membership}

\item{parms}{vector of fixed and random effects of length ncol(data\link{X}) + ncol(data\link{Z})}
}
\value{
The tibble given to \code{data} is returned with two additional columns. lp contains (XB + Zb) and A contains exp(XB + Zb)
}
\description{
calculate (XB + Zb) and exp(XB + Zb)
}
