% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{calcPenalty}
\alias{calcPenalty}
\title{calculate the penalty terms for dlp_b_b}
\usage{
calcPenalty(data, D, Z, wi, i)
}
\arguments{
\item{data}{tibble of data.}

\item{Z}{character vector of columns in data indicating cluster membership}

\item{D_theta}{diagonal matrix for given theta. Row and column names must
match Z columns in the data}
}
\description{
Calcualtes the penalty term for the second partial derivative of the pseudo-partial likelihood
with respect to the random effects, b.
}
