% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{calcRiskSets}
\alias{calcRiskSets}
\title{#' create a dataset for testing cox model stuff.
#'
#' takes three fixed effects and one parameter, theta, that represents
#' the variance of random effects drawn from a normal distribution.
#'}
\usage{
calcRiskSets(data, vars = NULL, varCol = NULL, A = A, index = index)
}
\arguments{
\item{data}{tibble containing all the data.}

\item{vars}{names of variable to multiply the risk set by. These variables will be pivoted longer. defaults to NULL.}

\item{varCol}{name to the product of vars and A.}

\item{A}{exponentiated linear predictor. i.e. exp(XB + Zb)}

\item{index}{numerical index for event times.}
}
\value{
Returns the tibble provided in data. If vars = NULL, the data will have the same number of rows, and an
additional column cumsum_A which is the cumulative sum of A, taken from the bottom of the data up. This is related
to the idea of risk sets.
}
\description{

}
\details{
n = control$k*control$nk

M = rep(1:control$k, each = control$nk)

X1 = rnorm(n, 0, 1)
X2 = rep(rnorm(control$k, 0, 1), each = control$nk)
}
\section{cluster level binary treatment allocation}{

}

