% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_populations_and_datasets.R
\name{check_limits}
\alias{check_limits}
\title{test if a value is in bounds.}
\usage{
check_limits(b, sqrtTheta)
}
\arguments{
\item{b}{a numeric vector}

\item{sqrtTheta}{a numeric scalar}
}
\description{
test if a value is in bounds.
}
