% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{check_var_labels}
\alias{check_var_labels}
\title{Check formula against dists names}
\usage{
check_var_labels(var_labels, dist_labels)
}
\arguments{
\item{var_labels}{Names of the variables in the right hand side of the formula}

\item{dist_labels}{Names of the expressions in dists}
}
\description{
Checks that the variables in the right hand side of the formula all have matching
expressions is the dists list.
}
