% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{control.list}
\alias{control.list}
\title{set control parameters for estimation loop}
\usage{
control.list(
  method = "ppl",
  grad = FALSE,
  re_only = TRUE,
  factr = 1e+07,
  reltol = sqrt(.Machine$double.eps),
  ndeps = 0.001,
  max_iter = 100
)
}
\description{
set control parameters for estimation loop
}
