% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{dlp_b}
\alias{dlp_b}
\title{partial derivative with respect to the random effects, b.}
\usage{
dlp_b(parms, X, cluster, stat_time, dij, D, data, ...)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{cluster}{charter vector with column name containing cluster membership}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}

\item{data}{tibble with all these columns}

\item{t}{column in data with time of failure/censoring data.}
}
\description{
partial derivative with respect to the random effects, b.
}
