% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{dlp_b_b}
\alias{dlp_b_b}
\title{Wang 2019 likelihood differentiated twice with respect to the random effects, b.}
\usage{
dlp_b_b(parms, X, Z, t, wi, wji, dij, i, D, data)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{Z}{character vector containing names of Z columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{wi}{column in data containing cluster-level weights.}

\item{wji}{column in data containing within-cluster weights.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}. row and column names must match \code{Z}}

\item{data}{tibble with all these columns}
}
\description{
Wang 2019 likelihood differentiated twice with respect to the random effects, b.
}
