% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{est_parameters}
\alias{est_parameters}
\title{estimate theta beta and b}
\usage{
est_parameters(
  formula,
  data,
  start_params = NULL,
  theta_start = NULL,
  control = control.list(),
  weights,
  cluster_weights
)
}
\description{
estimate theta beta and b
}
