% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{est_parameters_loop}
\alias{est_parameters_loop}
\title{estimate theta beta and b}
\usage{
est_parameters_loop(
  formula,
  data,
  method,
  start_params = NULL,
  theta_start = NULL,
  control = list(re_only = TRUE)
)
}
\description{
Instead of nested optim calls, uses alternating estimation of beta and theta until convergence.
}
