% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{est_theta}
\alias{est_theta}
\title{Shared frailty theta}
\usage{
est_theta(b, K_ppl)
}
\description{
In a shared frailty model with i.i.d. frailty terms, the solution to the
estimating equation for \theta has a closed form, implemented here.
}
