% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{est_theta2}
\alias{est_theta2}
\title{Shared frailty theta}
\usage{
est_theta2(par, b, K_ppl)
}
\description{
In a shared frailty model with i.i.d. frailty terms, the solution to the
estimating equation for \eqn{\theta} has a closed form, implemented here.
}
\details{
In the code, the identity matrix is the derivative of D, but for more
complex frailty models, D and it's derivative will be more complex (and this code won't work correctly).
}
