% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_functions.R
\name{estimate_all_parameters}
\alias{estimate_all_parameters}
\title{estimate_all_parameters}
\usage{
estimate_all_parameters(
  fixed_effects_start,
  random_effects_start,
  theta_start,
  x_columns,
  z_columns,
  t,
  i,
  wi,
  wji,
  dij,
  data,
  max_iterations = 1000,
  eps = .Machine$double.eps
)
}
\description{
estimates \eqn{\beta}, b and \eqn{\theta} using the method described in Wang
2019 and Ripatti et al 2001
}
