% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{estimate_parameters_loop}
\alias{estimate_parameters_loop}
\title{loops estimate_parameters and est_theta until convergence or max iterations
is reached.}
\usage{
estimate_parameters_loop(
  start_theta,
  start_parms,
  X,
  stat_time,
  cluster,
  dij,
  data,
  max_iter = 100,
  convergence_threshold = 1e-05
)
}
\description{
loops estimate_parameters and est_theta until convergence or max iterations
is reached.
}
