% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{fix_formula}
\alias{fix_formula}
\title{modifies a formula.}
\usage{
fix_formula(formula)
}
\description{
removes random effect terms
adds  + offset(offset)
intented to modify the formula slot in coxme call when fitting
coxph model with random effects as offsets.
}
