% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{get_group_label}
\alias{get_group_label}
\title{extract term.labels attribute}
\usage{
get_group_label(random_term)
}
\description{
\link{\code{terms}} returns the formula it is given with a decomposition of attributes attached.
This function is a wrapper for \code{attr(., "term.labels")}
}
