% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{get_start_theta}
\alias{get_start_theta}
\title{this will be a proper function that tries a few sensible point on the
theta sample space and returns the best one, based on theta_ipl.
Currently just returns a vector of ones of length n_theta}
\usage{
get_start_theta(n_theta)
}
\description{
this will be a proper function that tries a few sensible point on the
theta sample space and returns the best one, based on theta_ipl.
Currently just returns a vector of ones of length n_theta
}
