% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{K_prime_prime}
\alias{K_prime_prime}
\title{K''}
\usage{
K_prime_prime(parms, X, t, dij, theta, cluster, data, ...)
}
\description{
Ripatti Palmgren sub out K'' for K_ppl''. Here I implement K'' so I can compare the use of both
methods. This involves calculation of the breslow estimator of cumulative hazard.
}
