% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{lp}
\alias{lp}
\title{penalised partial likelihood in Ripatti 2000}
\usage{
lp(parms, X, t, dij, D, data, ...)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}

\item{data}{tibble with all these columns}

\item{Z}{character vector containing names of Z columns in data.}

\item{i}{column in data that identifies clusters.}
}
\description{
returns the PPL proposed by Ripatti and Palmgren
}
