% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{lp_grd}
\alias{lp_grd}
\title{lp_grd}
\usage{
lp_grd(parms, X, Z, t, wi, i, D, wji, dij, data)
}
\description{
combine the gradient function for \eqn{\beta} and b to make one vector of
gradients.
}
