% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_populations_and_datasets.R
\name{make_cluster_size}
\alias{make_cluster_size}
\title{calculate cluster size for a set of clusters}
\usage{
make_cluster_size(b, theta, base = 75, ... = ...)
}
\arguments{
\item{b}{vector of random numbers from N(0, \theta)}

\item{theta}{variance of the normal distribution that b is drawn from}

\item{base}{constant used is the calculation of cluster size

Calculates cluster sizes according to the method specified in Wang 2019.}
}
\description{
calculate cluster size for a set of clusters
}
