% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_populations_and_datasets.R
\name{make_population}
\alias{make_population}
\title{Make a finite population}
\usage{
make_population(
  theta,
  N,
  fixed_effects,
  random_effects = NULL,
  prCensoring,
  ...
)
}
\arguments{
\item{theta}{variance of the random effects}

\item{N}{number of clusters in the population}

\item{...}{arguments passed to other functions}
}
\value{
Returns a tibble with the columns
\itemize{
\item cluster factor with numeric levels identifying cluster membership
\item t failure or censoring time
\item d indicator of failure. 1 indicates failure, 0 censoring
\item X1:X3 fixed effect covariates. X1 ~ N(0, 1) and varies across individuals and
clusters. X2 ~ N(0, 1) varies across clusters, but constant within clusters.
\item X3 ~ binomial(0.5, N) Binary 'treatment allocation' at the cluster level.
\item Z1:ZN columns indicating cluster membership.
}

an additional attribute has been attached to the tibble: 'b', which holds the
cluster-level random effects/frailties
}
\description{
Make a finite population
}
\details{
cluster sizes are determined via make_cluster_size. Everything is set up
as in Wang 2019.
}
