% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{makeLong}
\alias{makeLong}
\title{helper for calcCrossProducts}
\usage{
makeLong(data, f)
}
\arguments{
\item{data}{A tibble containing an index column called index, and all columns named in f}

\item{f}{character vector of columns to make longer.}
}
\description{
takes a tibble and selects everything in the character vector f, as well
an index variable from sortAndIndex. pivots everything in f to long format,
and converts the 'name' column to a factor. Levels are determined by order they appear in the data.
}
