% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{one_dataset}
\alias{one_dataset}
\title{create a dataset for testing cox model stuff.}
\usage{
one_dataset(control)
}
\description{
takes three fixed effects and one parameter, theta, that represents
the variance of random effects drawn from a normal distribution.
}
