% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{rcpp_hello_world}
\usage{
rcpp_hello_world()
}
\arguments{
\item{no}{no params}
}
\description{
This function returns a list with character and numeric elements,  demonstrating the use of lists in C++, and testing that Rcpp is integrated properly into the package build process.
}
