% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_to_offset.R
\name{re_to_offset}
\alias{re_to_offset}
\title{turn a list of coxme random effects into a single offset column.}
\usage{
re_to_offset(data, model)
}
\description{
turn a list of coxme random effects into a single offset column.
}
