% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_populations_and_datasets.R
\name{sample_clusters}
\alias{sample_clusters}
\title{draw a sample of clusters and all members is those clusters}
\usage{
sample_clusters(population, cluster, n, ni, z_columns, prob = NULL)
}
\arguments{
\item{population}{tibble of the complete population to sample from}

\item{cluster}{variable in population identifying cluster membership}

\item{n}{number of clusters to sample}

\item{ni}{number of observations to sample from within each cluster}

\item{z_columns}{all the z columns in the population tibble}

\item{prop}{named vector of selection probabilities for clusters. Defaults to NULL i.e. simple
random sampling}
}
\description{
draw a sample of clusters and all members is those clusters
}
