% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{sortAndIndex}
\alias{sortAndIndex}
\title{sort and index data}
\usage{
sortAndIndex(data, sort_vars, index = "index")
}
\arguments{
\item{data}{A tibble to sort}

\item{index}{name to give the index variable. Default is "index".}

\item{sortVars}{tidyselected variables to sort by}
}
\description{
Sorts data by one or more variables, and adds an index to the sorted tibble
}
