% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{svycoxme.survey.design}
\alias{svycoxme.survey.design}
\title{svycoxme.survey.design}
\usage{
\method{svycoxme}{survey.design}(formula, design, subset = NULL, rescale = TRUE, ...)
}
\description{
check next release for the fix to the rescale issues.
}
