% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{svycoxme.svyrep.design}
\alias{svycoxme.svyrep.design}
\title{svycoxme}
\usage{
\method{svycoxme}{svyrep.design}(
  formula,
  design,
  subset = NULL,
  rescale = NULL,
  ...,
  control = coxme::coxme.control(),
  return.replicates = FALSE,
  na.action,
  multicore = getOption("survey.multicore")
)
}
\description{
function for rep weight designs
}
