% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{svycoxph.svyrep.design}
\alias{svycoxph.svyrep.design}
\title{svycoxph
overwrite survey:: method with my own to see if the way i'm assigning weights in the replicates is causing the error.}
\usage{
\method{svycoxph}{svyrep.design}(
  formula,
  design,
  subset = NULL,
  rescale = NULL,
  ...,
  return.replicates = FALSE,
  na.action,
  multicore = getOption("survey.multicore")
)
}
\description{
svycoxph
overwrite survey:: method with my own to see if the way i'm assigning weights in the replicates is causing the error.
}
