% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{theta_ipl}
\alias{theta_ipl}
\title{likelihood for theta}
\usage{
theta_ipl(
  theta,
  formula,
  parsed_data,
  other_args,
  weights,
  cluster_weights,
  mean_cluster_weight,
  rescaled_weights_used = FALSE
)
}
\description{
likelihood for theta
}
