% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{theta_ipl_fn_gr}
\alias{theta_ipl_fn_gr}
\title{Memoised likelihood and gradient, for use in optimisation.}
\usage{
theta_ipl_fn_gr(formula, parsed_data, other_args, weights, cluster_weights)
}
\description{
Memoised likelihood and gradient, for use in optimisation.
}
