% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{theta_ipl_gr}
\alias{theta_ipl_gr}
\title{first derivative of the likelihood for theta}
\usage{
theta_ipl_gr(one_theta, parms, X, stat_time, dij, cluster, data)
}
\description{
This is the simplified version, assuming shared frailty.
}
