% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{theta_ipl_gr2}
\alias{theta_ipl_gr2}
\title{gradient for theta likelihood}
\usage{
theta_ipl_gr2(theta, formula, parsed_data, other_args, ests)
}
\description{
takes beta and b as arguments, instead of internally optimising them.
}
