% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{theta_ipl_gr3}
\alias{theta_ipl_gr3}
\title{calculate gradient for theta_ipl using K prime prime, instead of subbing in K_ppl prime prime.}
\usage{
theta_ipl_gr3(theta, formula, parsed_data, other_args)
}
\description{
calculate gradient for theta_ipl using K prime prime, instead of subbing in K_ppl prime prime.
}
