% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{control.list}
\alias{control.list}
\title{set control parameters for estimation loop}
\usage{
control.list(
  method = "ppl",
  grad = FALSE,
  re_only = TRUE,
  factr = 1000,
  reltol = 1e-05,
  ndeps = 0.001,
  max_iter = 100
)
}
\description{
defaults change factr to 1e3 from 1e7, and reltol from sqrt(.Machine$double.eps) to 1e-13
this makes the run-to-run variation smaller (negligible), but increases time to convergence.
}
