% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{dlp_beta}
\alias{dlp_beta}
\title{partial derivative with respect to the fixed effects, \eqn{\beta}.}
\usage{
dlp_beta(parms, X, Z, t, wi, wji, dij, data)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{Z}{character vector containing names of Z columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{wi}{column in data containing cluster-level weights.}

\item{wji}{column in data containing within-cluster weights.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{data}{tibble with all these columns}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}
}
\description{
partial derivative with respect to the fixed effects, \eqn{\beta}.
}
