% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{draw_event_times}
\alias{draw_event_times}
\title{Create a dataset of single or recurrent events for subjects with possibly time-varying covariate data.}
\usage{
draw_event_times(
  formula,
  data,
  coefficients = c(),
  random_effect_variance = list(),
  id,
  baseline_hazard = 1,
  t = 0,
  event = c("single", "recurrent"),
  origin = 0
)
}
\arguments{
\item{formula}{Model equation for the data generation.}

\item{data}{Covariates in the \code{formula}.}

\item{coefficients}{Named vector of fixed effect coefficients.}

\item{random_effect_variance}{Named vector of random effects. Names must match random effect formula terms.}

\item{id}{The data column that identifies observations belonging to one subject.}

\item{baseline_hazard}{scalar or vector with stepwise hazards over time}

\item{t}{change points for \code{baseline_hazard}}

\item{event}{option for a single event or recurrent events per subject}

\item{origin}{when the time scale starts}
}
\description{
Create a dataset of single or recurrent events for subjects with possibly time-varying covariate data.
}
