% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{estimate all parameters in a shared frailty model}
\usage{
estimate_parameters(start_parms, theta, X, t, cluster, dij, data)
}
\description{
Algorithm iterates between estimating fixed and random effects for a given
theta, and estimating theta given the current estimates fixed and random effects,
until convergence is reached.
}
