% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{estimate_parameters_loop}
\alias{estimate_parameters_loop}
\title{loops estimate_parameters and est_theta until convergence or max iterations
is reached.}
\usage{
estimate_parameters_loop(
  beta,
  theta,
  k,
  nk,
  cluster = "M",
  max_iter = 100,
  convergence_threshold = 1e-05,
  start_theta = 4,
  X = c("X1", "X2", "X3")
)
}
\description{
loops estimate_parameters and est_theta until convergence or max iterations
is reached.
}
