% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang_likelihoods.R
\name{lp}
\alias{lp}
\title{pseudo likelihood in Wang 2019}
\usage{
lp(parms, X, Z, t, i, wi, wji, dij, D, data)
}
\arguments{
\item{parms}{numeric vector of \eqn{\beta} and b values i.e \code{c(beta, b)}. The condition \code{length(parms) == length(c(X, Z))} must be satisfied.}

\item{X}{character vector containing names of X columns in data.}

\item{Z}{character vector containing names of Z columns in data.}

\item{t}{column in data with time of failure/censoring data.}

\item{i}{column in data that identifies clusters.}

\item{wi}{column in data containing cluster-level weights.}

\item{wji}{column in data containing within-cluster weights.}

\item{dij}{column in data indicating if observation was censored or a failure.}

\item{D}{diagonal matrix with given \eqn{\theta}. Must be a square matrix of order \code{length(Z)}.}

\item{data}{tibble with all these columns}
}
\description{
returns the likelihood proposed in Wang 2019
}
