% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripatti_likelihoods.R
\name{lp_gr_beta}
\alias{lp_gr_beta}
\alias{lp_gr_beta.ppl}
\alias{lp_gr_b}
\alias{lp_gr_b.ppl}
\alias{lp_gr}
\alias{lp_gr.ppl}
\title{lp_gr_beta}
\usage{
lp_gr_beta(params, formula, parsed_data, other_args)

\method{lp_gr_beta}{ppl}(params, formula, parsed_data, other_args)

lp_gr_b(params, formula, parsed_data, other_args)

\method{lp_gr_b}{ppl}(params, formula, parsed_data, other_args)

lp_gr(params, formula, parsed_data, other_args)

\method{lp_gr}{ppl}(params, formula, parsed_data, other_args)
}
\description{
caluculate jacobian
}
