% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycoxme.R
\name{residuals.coxme}
\alias{residuals.coxme}
\title{residuals.coxme
calculate residuals for a coxme model}
\usage{
\method{residuals}{coxme}(
  object,
  data,
  weighted = TRUE,
  include_re = FALSE,
  type = c("score", "dfbeta", "dfbetas"),
  ...
)
}
\description{
residuals.coxme
calculate residuals for a coxme model
}
