% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{residuals.coxme}
\alias{residuals.coxme}
\title{residuals.coxme
calculate residuals for a coxme model}
\usage{
\method{residuals}{coxme}(
  object,
  data,
  type = c("martingale", "deviance", "score", "schoenfeld", "dfbeta", "dfbetas",
    "scaledsch", "partial"),
  ...
)
}
\description{
residuals.coxme
calculate residuals for a coxme model
}
