% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycoxme.R
\name{svycoxme}
\alias{svycoxme}
\title{Survey-weighted mixed-effects Cox models}
\usage{
svycoxme(formula, design, subset = NULL, rescale = TRUE, ...)
}
\arguments{
\item{formula}{Model formula.}

\item{design}{\code{survey.design} object. It must contain all variables in the formula.}

\item{subset}{Expression to select a subpopulation.}

\item{rescale}{Rescale weights to improve numerical stability.}

\item{...}{Other arguments passed to \link[coxme]{coxme}.}
}
\value{
An object of class \code{svycoxme}.
}
\description{
Fit a mixed-effect proportional hazards model to data from a complex design.
}
