% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R
\name{svycoxme}
\alias{svycoxme}
\title{functions needed to make svycoxme go.}
\usage{
svycoxme(formula, design, subset = NULL, ...)
}
\description{
svycoxme
}
\details{
Need a function that takes the random effect information from coxme and converts it into an offset.
This function will need to handle interaction terms from the random effects.
Generic method
}
