% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycoxme.R
\docType{package}
\name{svycoxme}
\alias{-package}
\alias{svycoxme}
\title{svycoxme: A package implementing mixed effect proportional hazard models for
complex samples}
\usage{
svycoxme(formula, design, subset = NULL, ...)
}
\arguments{
\item{formula}{Model formula.}

\item{design}{\code{survey.design} object. It must contain all variables in the formula.}

\item{subset}{Expression to select a subpopulation.}

\item{...}{Other arguments passed to \code{coxme}.}

\item{rescale}{Rescale weights to improve numerical stability.}
}
\value{
An object of class \code{svycoxme}.
}
\description{
The svycoxme package allows users to fit \code{coxme} models with survey
designs from \code{survey}.
}
\section{Mypackage functions}{

The mypackage functions ...
}

