% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxme_survey.R, R/svycoxme.R
\docType{package}
\name{svycoxme}
\alias{svycoxme-package}
\alias{svycoxme}
\title{functions needed to make svycoxme go.}
\usage{
svycoxme(formula, design, subset = NULL, ...)
}
\description{
svycoxme

The svycoxme package allows users to fit \code{coxme} models with survey
designs from \code{survey}.
}
\details{
Need a function that takes the random effect information from coxme and converts it into an offset.
This function will need to handle interaction terms from the random effects.
Generic method

rescale should always be true in my opinion. Coxme often fails with unscaled weights. When it runs, the estimates it produces are really wrong.
}
\section{Mypackage functions}{

The mypackage functions ...
}

\author{
\strong{Maintainer}: Bradley Drayton \email{brad.drayton@auckland.ac.nz}

}
